% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compte_grp_lin}
\alias{compte_grp_lin}
\title{Compter les groupes de 1 consecutifs lignes}
\usage{
compte_grp_lin(grille, num)
}
\arguments{
\item{grille}{La grille de jeu}

\item{num}{Le numero de ligne}
}
\value{
Un vecteur avec le nombre de groupes de 1 consecutifs
}
\description{
Fonction pour compter les groupes de 1 consecutifs dans une ligne de la grille
}
\examples{
grille_exemple <- matrix(c(1, 0, 1, 1, 0, 1, 0, 1, 1), nrow = 3, byrow = TRUE)
compte_grp_lin(grille_exemple, 2)
}
