% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{difficulty}
\alias{difficulty}
\title{Fonction traduire une difficulte}
\usage{
difficulty(str)
}
\arguments{
\item{str}{La difficulte choisie: "Facile", "Moyen", "Difficile", "Très difficile"}
}
\value{
Un pourcentage de cases noirs pour une future grille
}
\description{
Sert a traduire une difficulte en une proportion
}
\examples{
difficulty("Facile")
}
