% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grille}
\alias{grille}
\title{Generer une grille aleatoire}
\usage{
grille(taille, difficulte)
}
\arguments{
\item{taille}{La taille de la grille}

\item{difficulte}{La proportion de cases noires}
}
\value{
Une grille aleatoire
}
\description{
Genere une grille aleatoire avec une proportion de cases noirs dependant de la difficulte choisie
}
\examples{
grille(5,0.3)
}
