% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Fonction du serveur pour l'application Shiny Picross}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Les valeurs des widgets d'entree de l'application.}

\item{output}{Les sorties à afficher dans l'interface utilisateur.}

\item{session}{Une session Shiny.}
}
\description{
Cette fonction definit le comportement du serveur pour l'application Shiny Picross.
}
