% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Fonction pour creer l'interface utilisateur Shiny}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
ui
}
\value{
Une interface utilisateur Shiny.
}
\description{
Cette fonction cree une interface utilisateur Shiny avec un titre et une barre latérale.
}
\keyword{datasets}
