% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vectcol}
\alias{vectcol}
\title{Fonction applique compte_grp_col a chaque colonne de la grille.}
\usage{
vectcol(grille)
}
\arguments{
\item{grille}{La grille de jeu}
}
\value{
Une liste de vecteurs où chaque element represente les groupes de 1 consecutifs dans une colonne.
}
\description{
Fonction applique compte_grp_col a chaque colonne de la grille.
}
\examples{
grille_exemple <- matrix(c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1), nrow = 5, byrow = TRUE)
vectcol(grille_exemple)
}
