% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verif}
\alias{verif}
\title{Fonction de verification}
\usage{
verif(cellStates, grille)
}
\arguments{
\item{cellStates}{L'etat des cellules}

\item{grille}{La grille de jeu}
}
\value{
TRUE si la grille est correctement resolue, sinon FALSE
}
\description{
Fonction pour verifier si la grille est resolue correctement, fonctionne si la grille n'est pas a solution unique.
}
\examples{
# Creer une grille de jeu de 3x3 avec une solution correcte
grille_correcte <- matrix(c(1, 0, 1,
                            0, 1, 0,
                            1, 0, 1), nrow = 3, byrow = TRUE)

# Definir l'etat des cellules
cell_states <- '{"cell_1_1":1, "cell_1_2":0, "cell_1_3":1,
                 "cell_2_1":0, "cell_2_2":1, "cell_2_3":0,
                 "cell_3_1":1, "cell_3_2":0, "cell_3_3":1}'

# Verifier si la grille est correctement resolue
verif(cell_states, grille_correcte)
}
